# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Documentação da API.",
        "title": "Retro Board API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:9000",
    "basePath": "/v1",
    "paths": {
...
        "/teams/:teamId/boards/:boardId": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Updates the name or activation status of a board.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Board"
                ],
                "summary": "Edit a board",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Team ID",
                        "name": "teamId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Board ID",
                        "name": "boardId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Board update details",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.EditBoardRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Board successfully updated"
                    },
                    "400": {
                        "description": "Invalid request (e.g., missing parameters or validation error)",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorEnvelope"
                        }
                    },
                    "404": {
                        "description": "Team or board not found",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorEnvelope"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorEnvelope"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Deletes a board associated with the specified team and board ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Board"
                ],
                "summary": "Delete a board",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Team ID",
                        "name": "teamId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Board ID",
                        "name": "boardId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Board successfully deleted"
                    },
                    "400": {
                        "description": "Invalid request (e.g., missing parameters or validation error)",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorEnvelope"
                        }
                    },
                    "404": {
                        "description": "Team or board not found",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorEnvelope"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorEnvelope"
                        }
                    }
                }
            }
        },
...
        "/teams/:teamId/boards/:boardId/columns/:columnId/cards/:cardId": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves the details of a specific card based on its ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Board"
                ],
                "summary": "Get a single card by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Team ID",
                        "name": "teamId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Board ID",
                        "name": "boardId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Column ID",
                        "name": "columnId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Card ID",
                        "name": "cardId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Card details",
                        "schema": {
                            "$ref": "#/definitions/board.ListCardEnvelop"
                        }
                    },
                    "400": {
                        "description": "Invalid request (e.g., missing parameters or invalid input data)",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorEnvelope"
                        }
                    },
                    "404": {
                        "description": "Card not found",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorEnvelope"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorEnvelope"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Updates the text content of a card.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Board"
                ],
                "summary": "Edit a card",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Team ID",
                        "name": "teamId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Board ID",
                        "name": "boardId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Column ID",
                        "name": "columnId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Card ID",
                        "name": "cardId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Card update details",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.EditCardRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Card successfully updated"
                    },
                    "400": {
                        "description": "Invalid request (e.g., missing parameters or validation error)",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorEnvelope"
                        }
                    },
                    "404": {
                        "description": "Team, board, column, or card not found",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorEnvelope"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorEnvelope"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Deletes a card associated with the specified team, board, column, and card ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Board"
                ],
                "summary": "Delete a card",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Team ID",
                        "name": "teamId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Board ID",
                        "name": "boardId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Column ID",
                        "name": "columnId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Card ID",
                        "name": "cardId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Card successfully deleted"
                    },
                    "400": {
                        "description": "Invalid request (e.g., missing parameters or validation error)",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorEnvelope"
                        }
                    },
                    "404": {
                        "description": "Team, board, column, or card not found",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorEnvelope"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorEnvelope"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "board.ListCardEnvelop": {
            "type": "object",
            "properties": {
                "card": {
                    "$ref": "#/definitions/dtos.ListCardResponse"
                }
            }
        },
...
        "dtos.EditBoardRequest": {
            "type": "object",
            "properties": {
                "active": {
                    "type": "boolean",
                    "example": true
                },
                "name": {
                    "type": "string",
                    "example": "new-board1"
                }
            }
        },
        "dtos.EditCardRequest": {
            "type": "object",
            "properties": {
                "text": {
                    "type": "string",
                    "example": "new-card1-text"
                }
            }
        },
...
        "dtos.ListCardResponse": {
            "type": "object",
            "properties": {
                "column_id": {
                    "type": "integer",
                    "example": 2
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "member_id": {
                    "type": "integer",
                    "example": 2
                },
                "text": {
                    "type": "string",
                    "example": "card1-text"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "utils.ErrorEnvelope": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "error message"
                }
            }
        }
    },
    "securityDefinitions": {
        "": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}